+++ atacseq/ccbr_trim_align/install_R_packages2.R
@@ -1,0 +1,2 @@
+
+install.packages("plotly", repos="http://cran.us.r-project.org")